test: split JreCipherWrapperTest
